+++ presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractArrayBlock.java
@@ -90,7 +90,7 @@ public abstract class AbstractArrayBlock
     }
 
     @Override
-    public int getRegionSizeInBytes(int position, int length)
+    public long getRegionSizeInBytes(int position, int length)
     {
         int positionCount = getPositionCount();
         if (position < 0 || length < 0 || position + length > positionCount) {
